app/requirements.txt
flask==2.3.3
gunicorn==21.2.0
google-generativeai==0.3.0
pathlib==1.0.1
python-dotenv==1.0.0